test: extract overlapping-outputs regression to dedicated class
Keep ModuleGraphPluginFunctionalTest under detekt LargeClass after
the issue #72 functional coverage was added.
